requirements.txt
fastapi
uvicorn
python-jose
python-dotenv
mysql-connector-python
pydantic